test: split railway unit/integration
